+++ CHANGELOG.md
@@ -6,4 +6,18 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
-- Initial release
+## [0.2.0] - 2025-05-06
+- Added ability to select target files for building
+- Added condition "when Ruby source file is open" to file transfer on save
+
+## added
+
+## [0.1.5] - 2025-03-27
+
+### Added
+
+### Changed
+- CRC calculation method fixed
+
+### Note
+- Last Test Version